requirements.txt
APScheduler==3.9.1
apturl==0.5.2
asgiref==3.6.0
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
coverage==6.5.0
cryptography==39.0.1
cupshelpers==1.0
daphne==4.0.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.4
django-allauth==0.51.0
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-environ==0.9.0
django-filter==23.3
django-oauth-toolkit==2.2.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-restframework==0.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.2.1
duplicity==0.8.21
environ==1.0
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
jeepney==0.7.1
jwcrypto==1.4.2
kazam==1.4.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.28.2
requests-oauthlib==1.3.1
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stripe==5.0.0
stripe3==0.0.0
systemd-python==234
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
wrapt==1.14.1
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.5.2
